refactor: calculateCouponPrice 폴더 이동
- utils보다 도메인 관련 로직이니 domain폴더로 이동
