+++ src/domain/calculateCouponPrice.ts
