+++ src/domain/getCouponUIData.ts
@@ -1,5 +1,6 @@
-import { calculateCouponPrice, isCouponDisabled } from '../utils';
+import { isCouponDisabled } from '../utils';
 import { Coupon, CartItemTypes } from '../types';
+import { calculateCouponPrice } from './calculateCouponPrice';
 
 interface useCouponUIProps {
   coupons: Coupon[];
