+++ src/domain/index.ts
@@ -1,2 +1,3 @@
+export * from './calculateCouponPrice';
 export * from './getCouponUIData';
 export * from './getOrderCalculationData';
